+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/dm/service/NotificationService.java
@@ -6,8 +6,8 @@ import com.capstone.notificationservice.domain.dm.dto.DmNotificationDto;
 import com.capstone.notificationservice.domain.dm.dto.response.DmNotificationResponse;
 import com.capstone.notificationservice.domain.dm.entity.DmNotification;
 import com.capstone.notificationservice.domain.dm.exception.DmException;
-import com.capstone.notificationservice.domain.dm.respository.EmitterRepository;
-import com.capstone.notificationservice.domain.dm.respository.NotificationRepository;
+import com.capstone.notificationservice.domain.dm.repository.EmitterRepository;
+import com.capstone.notificationservice.domain.dm.repository.NotificationRepository;
 import com.capstone.notificationservice.domain.user.entity.User;
 import com.capstone.notificationservice.domain.user.service.UserService;
 import com.capstone.notificationservice.global.exception.Code;
